a lot of changes done

+++ app.js
@@ -1,3 +1,7 @@
+if(process.env.NODE_ENV !== 'production') {
+    require('dotenv').config()
+}
+
 //imports
 const express = require('express')
 const app = express()
@@ -10,11 +14,17 @@ const flash = require('connect-flash')
 const passport = require('passport')
 const passportLocal = require('passport-local')
 
+//cloudinary
+const multer = require('multer')
+const { cloudinary, storage } = require('./cloudinary/config')
+const upload = multer({ storage })
+
 const Dog = require('./models/dogs')
 const User = require('./models/user')
 
 const catchAsync = require('./public/javascript/CatchAsync')
 const ExpressError = require('./public/javascript/ExpressError')
+const isLoggedIn = require('./public/javascript/middleware')
 
 const validateDog = require('./public/javascript/Joi')
 
@@ -59,11 +69,6 @@ app.use(session( sessionConfig))
 //flash
 app.use(flash())
 
-app.use((req,res,next)=>{
-    res.locals.success = req.flash('success')
-    res.locals.error = req.flash('error')
-    next()
-})
 
 //passport authentication
 app.use(passport.initialize())
@@ -73,37 +78,53 @@ passport.use(new passportLocal(User.authenticate()))
 passport.serializeUser(User.serializeUser())
 passport.deserializeUser(User.deserializeUser())
 
-//routes
+
+app.use((req,res,next)=>{
+    if(!['/login', '/'].includes(req.originalUrl)) {
+        req.session.returnTo = req.originalUrl;
+    }
+    res.locals.currentUser = req.user
+    res.locals.success = req.flash('success')
+    res.locals.error = req.flash('error')
+    next()
+})
+
+
+//ROUTES
 app.get('/', (req,res)=>{
     res.render('home')
 })
 
-//show all dogs
+//SHOW ALL DOGS
 app.get('/dogs', catchAsync(async(req,res) => {
     const dogs = await Dog.find({})
+
     res.render('dogs/index', { dogs })
 }))
 
 //POST NEW DOG
-app.get('/new', (req,res)=>{
-    if(!req.isAuthenticated()){
-        req.flash('error', 'You must sign in first')
-        res.redirect('/login')
-    }
+app.get('/dogs/new', isLoggedIn, (req,res)=>{
     res.render('dogs/new')
 })
 
-app.post('/dogs', validateDog, catchAsync(async(req,res, next)=>{
+app.post('/dogs', isLoggedIn, upload.array('dog[image]'), validateDog, catchAsync(async(req,res, next)=>{
     const dog = new Dog(req.body.dog)
+    dog.author = req.user._id
+    dog.images = req.files.map(f => ({url: f.path, filename: f.filename}))
     await dog.save()
     req.flash('success', 'Successfully published your puppy')
     res.redirect(`/dogs/${dog._id}`)
 }))
 
+
+app.post('/dogs', upload.array('dog[image]'), (req,res)=>{
+    console.log(req.body, req.files);
+    res.send('it woek')
+})
 //SHOW SINGLE DOG
 app.get('/dogs/:id', catchAsync(async(req,res) => {
     const {id} = req.params
-    const dog = await Dog.findById(id)
+    const dog = await Dog.findById(id).populate('author')
     if(!dog){
         req.flash('error', 'Thankfuly the puppy already found a new home. Check out some others in need of love.')
         return res.redirect('/dogs')
@@ -112,7 +133,7 @@ app.get('/dogs/:id', catchAsync(async(req,res) => {
 }))
 
 //EDIT SINGLE DOG
-app.get('/dogs/:id/edit', catchAsync(async(req,res) => {
+app.get('/dogs/:id/edit', isLoggedIn, catchAsync(async(req,res) => {
     const {id} = req.params
     const dog = await Dog.findById(id)
     if(!dog){
@@ -122,18 +143,27 @@ app.get('/dogs/:id/edit', catchAsync(async(req,res) => {
     res.render('dogs/edit', { dog })
 }))
 
-app.put('/dogs/:id', validateDog, catchAsync(async(req,res)=>{ 
+app.put('/dogs/:id', isLoggedIn, validateDog, catchAsync(async(req,res)=>{ 
     if(!req.body.dog) throw new ExpressError('Invalid try buddy', 400)
     const {id} = req.params
+    const dog = await Dog.findById(id)
+    if(!dog.author.equals(req.user._id)){
+        req.flash('error', 'Nice try buddy')
+        return res.redirect('/dogs')
+    }
     const updatedDog = await Dog.findByIdAndUpdate(id, {...req.body.dog})
-    updatedDog.save()
     req.flash('success', 'Successfully updated your puppys information')
     res.redirect(`/dogs/${updatedDog._id}`)
 }))
 
 //delete a doggie = succesfully adopted
-app.delete('/dogs/:id', catchAsync(async(req,res)=>{
+app.delete('/dogs/:id', isLoggedIn, catchAsync(async(req,res)=>{
     const {id} = req.params
+    const dog = await Dog.findById(id)
+    if(!dog.author.equals(req.user._id)){
+        req.flash('error', 'Nice try buddy')
+        return res.redirect('/dogs')
+    }
     await Dog.findByIdAndDelete(id)
     req.flash('success', 'Thank you for getting a puppy a new home!')
     res.redirect('/dogs') 
@@ -148,9 +178,13 @@ app.post('/register', catchAsync(async(req,res)=>{
     try {
     const {email,username,password} = req.body 
     const user = new User({email, username})
-    await User.register(user, password)
-    req.flash('success', 'Wellcome to Posvoji.si')
-    res.redirect('/')    
+    const registeredUser = await User.register(user, password)
+    req.login(registeredUser, (err) => {
+        if(err) return next(err)
+        req.flash('success', 'Wellcome to Posvoji.si')
+        res.redirect('/') 
+    })
+       
 } catch(e){
         req.flash('error', e.message)
         res.redirect('/register')
@@ -163,6 +197,14 @@ app.get('/login', (req,res)=>{
 
 app.post('/login', passport.authenticate('local', {failureFlash: true, failureRedirect: '/login'}), (req, res)=>{
     req.flash('success', 'Welcome back!')
+    const redirectUrl =  req.session.returnTo || '/'
+    delete req.session.returnTo
+    res.redirect(redirectUrl)
+})
+
+app.get('/logout', (req,res)=>{
+    req.logout()
+    req.flash('success', 'Goodbye')
     res.redirect('/')
 })
